fix: handle TV menu and back navigation
